# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobAcceptedResponse": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "operation": {
            "title": "Operation",
            "type": "string"
          },
          "status": {
            "default": "queued",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "operation"
        ],
        "title": "JobAcceptedResponse",
        "type": "object"
      },
...
      "StudyPlanGenerateRequest": {
        "properties": {
          "gap_ratio": {
            "default": 0.4,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Gap Ratio",
            "type": "number"
          }
        },
        "title": "StudyPlanGenerateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "AI-powered adaptive learning platform — Grade R to 7. CAPS-aligned. POPIA-compliant.",
    "title": "EduBoost SA V2",
    "version": "1.0.0-rc1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/study-plans/generate/{learner_id}": {
      "post": {
        "operationId": "generate_study_plan_api_v2_study_plans_generate__learner_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "learner_id",
            "required": true,
            "schema": {
              "title": "Learner Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyPlanGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Generate Study Plan",
        "tags": [
          "V2 Study Plans"
        ]
      }
    },
    "/api/v2/study-plans/{learner_id}": {
      "post": {
        "operationId": "generate_study_plan_api_v2_study_plans__learner_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "learner_id",
            "required": true,
            "schema": {
              "title": "Learner Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyPlanGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Generate Study Plan",
        "tags": [
          "V2 Study Plans"
        ]
      }
    },
...
    "/v2/study-plans/generate/{learner_id}": {
      "post": {
        "operationId": "generate_study_plan_v2_study_plans_generate__learner_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "learner_id",
            "required": true,
            "schema": {
              "title": "Learner Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyPlanGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Generate Study Plan",
        "tags": [
          "V2 Study Plans"
        ]
      }
    },
    "/v2/study-plans/{learner_id}": {
      "post": {
        "operationId": "generate_study_plan_v2_study_plans__learner_id__post",
        "parameters": [
          {
            "in": "path",
            "name": "learner_id",
            "required": true,
            "schema": {
              "title": "Learner Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyPlanGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAcceptedResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Generate Study Plan",
        "tags": [
          "V2 Study Plans"
        ]
      }
    },
...
  },
...
}